backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-pptx>=0.6.21
httpx>=0.26.0
python-dotenv>=1.0.0
